FormatOps: move traversal methods to FormatTokens
